+++ tests/e2e/ingress_requirer.py
@@ -1,3 +1,6 @@
+# Copyright 2026 Canonical Ltd.
+# See LICENSE file for licensing details.
+
 """Ingress requirer any charm."""
 
 import pathlib
